refactor EntityTreeItem to its own file
